pkg/cover: split into ELF-dependent/independent parts
